remove old zip and upload code split file service into object for mockability
